feat: improve on benchmarking
Add --expiration and --capacity flags for the
benchmarking tool to allow for finer control
and testing of the storage layer options.
The expiration flag allows for changing the global
time to live value in the storage layer.
The capacity flag allows for changing the maximum
capacity of the storage layer.